Add files via upload
This is the first upload, I have been working on this for a few days. Still fleshing things out and writing lots of comments. Not really sure if I am doing it correctly, so things are a bit unstable and subject to change a lot. Trying to avoid making it too large, and still a beginner programmer so my techniques are still basic. But none the less I wanted to start tracking the changes on github publicly as part of building a portfolio